# api/openapi.yaml
openapi: 3.0.3
info:
  title: Shop Public API
  version: 1.4.0
  description: |
    Public and back-office API of the demo shop platform: catalog, carts, orders,
    payments, shipments, reviews, promo codes and the user back office.

    The document is maintained by hand and is the contract of record for the
    `shop-api-spring` service. It is intentionally NOT generated from the running
    application.
  contact:
    name: Shop Platform Team
    email: platform@shop.example.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/shop
    description: Local developer instance (server.servlet.context-path = /shop)
  - url: https://api.shop.example.com/shop
    description: Production
...
security:
  - bearerAuth: []
paths:
  /api/v1/products:
    get:
      operationId: listProducts
      summary: List catalog products
      tags: [catalog]
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: category
          in: query
          required: false
          description: Filter by product category
          schema:
            $ref: '#/components/schemas/ProductCategory'
        - name: q
          in: query
          required: false
          description: Case-insensitive substring match against the product title
          schema:
            type: string
            maxLength: 120
      responses:
        '200':
          description: Page of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createProduct
      summary: Create a catalog product
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created
          headers:
            Location:
              description: URI of the created product
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products/{id}:
    parameters:
      - $ref: '#/components/parameters/ProductIdParam'
    get:
      operationId: getProduct
      summary: Get a single product
      tags: [catalog]
      security: []
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replaceProduct
      summary: Replace a product
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteProduct
      summary: Delete a product
      tags: [catalog]
      responses:
        '204':
          description: Product deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/payments:
    get:
      operationId: listPayments
      summary: List payments
      tags: [payments]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - name: orderId
          in: query
          required: false
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: Page of payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: authorizePayment
      summary: Authorize a payment for an order
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    options:
      operationId: describePaymentOptions
      summary: Discover supported payment operations and methods
      tags: [payments]
      security: []
      responses:
        '200':
          description: Supported operations
          headers:
            Allow:
              description: Allowed HTTP methods
              schema:
                type: string
            X-Payment-Methods:
              description: Comma separated list of supported payment method kinds
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/payments/{id}:
    get:
      operationId: getPayment
      summary: Get a single payment
      tags: [payments]
      parameters:
        - $ref: '#/components/parameters/PaymentIdParam'
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/payments/{id}/refund:
    post:
      operationId: refundPayment
      summary: Refund a captured or authorized payment
      tags: [payments]
      parameters:
        - $ref: '#/components/parameters/PaymentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Payment refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/promo-codes:
    get:
      operationId: listPromoCodes
      summary: List promo codes
      tags: [promo]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of promo codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createPromoCode
      summary: Create a promo code
      tags: [promo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreateRequest'
      responses:
        '201':
          description: Promo code created
          headers:
            Location:
              description: URI of the created promo code
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/promo-codes/{code}/redeem:
    post:
      operationId: redeemPromoCode
      summary: Redeem a promo code for an order
      tags: [promo]
      security: []
      parameters:
        - name: code
          in: path
          required: true
          description: Human readable promo code
          schema:
            type: string
            pattern: '^[A-Z0-9]{4,16}$'
        - name: orderId
          in: query
          required: true
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: Redemption result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/reviews:
    get:
      operationId: listReviews
      summary: List product reviews
      tags: [reviews]
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: productId
          in: query
          required: false
          schema:
            type: string
            maxLength: 40
        - name: minRating
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: Page of reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createReview
      summary: Publish a new product review
      tags: [reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      responses:
        '201':
          description: Review accepted for moderation
          headers:
            Location:
              description: URI of the created review
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/reviews/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Review identifier
        schema:
          type: string
          maxLength: 40
    get:
      operationId: getReview
      summary: Get a single review
      tags: [reviews]
      security: []
      responses:
        '200':
          description: Review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteReview
      summary: Delete a review
      tags: [reviews]
      responses:
        '204':
          description: Review deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Customer or operator access token issued by the shop identity provider
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Static back-office key, required for every /admin operation
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Zero based page index
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Page size
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 20
    ProductIdParam:
      name: id
      in: path
      required: true
      description: Product identifier
      schema:
        type: string
        maxLength: 40
...
    PaymentIdParam:
      name: id
      in: path
      required: true
      description: Payment identifier
      schema:
        type: string
        maxLength: 40
...
  responses:
    BadRequest:
      description: The request payload or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller is authenticated but not allowed to perform the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The operation conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: RUB
...
    ProductCategory:
      type: string
      enum: [ELECTRONICS, BOOKS, CLOTHING, GROCERY, HOME]
    Product:
      type: object
      required: [id, sku, title, category, price, stock]
      properties:
        id:
          type: string
        sku:
          type: string
          pattern: '^SKU-[0-9]{5}$'
        title:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          $ref: '#/components/schemas/Money'
        stock:
          type: integer
          format: int32
          minimum: 0
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 24
        updatedAt:
          type: string
          format: date-time
    ProductRequest:
      type: object
      required: [sku, title, category, price]
      properties:
        sku:
          type: string
          pattern: '^SKU-[0-9]{5}$'
        title:
          type: string
          minLength: 3
          maxLength: 120
        description:
          type: string
          maxLength: 2000
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          $ref: '#/components/schemas/Money'
        stock:
          type: integer
          format: int32
          minimum: 0
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 24
    ProductPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Product'
...
    PaymentStatus:
      type: string
      enum: [PENDING, AUTHORIZED, CAPTURED, REFUNDED, DECLINED]
    PaymentMethod:
      type: object
      description: Polymorphic payment instrument, discriminated by the `kind` property.
      required: [kind]
      properties:
        kind:
          type: string
          enum: [CARD, WALLET]
      discriminator:
        propertyName: kind
        mapping:
          CARD: '#/components/schemas/CardPaymentMethod'
          WALLET: '#/components/schemas/WalletPaymentMethod'
      oneOf:
        - $ref: '#/components/schemas/CardPaymentMethod'
        - $ref: '#/components/schemas/WalletPaymentMethod'
    CardPaymentMethod:
      type: object
      required: [kind, last4, expiry, brand]
      properties:
        kind:
          type: string
          enum: [CARD]
        last4:
          type: string
          pattern: '^[0-9]{4}$'
        expiry:
          type: string
          pattern: '^(0[1-9]|1[0-2])/[0-9]{2}$'
        brand:
          type: string
          enum: [VISA, MASTERCARD, MIR]
    WalletPaymentMethod:
      type: object
      required: [kind, provider, walletId]
      properties:
        kind:
          type: string
          enum: [WALLET]
        provider:
          type: string
          maxLength: 40
        walletId:
          type: string
          maxLength: 64
    Payment:
      type: object
      required: [id, orderId, status, amount, method]
      properties:
        id:
          type: string
        orderId:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        amount:
          $ref: '#/components/schemas/Money'
        method:
          $ref: '#/components/schemas/PaymentMethod'
        providerReference:
          type: string
        createdAt:
          type: string
          format: date-time
    PaymentRequest:
      type: object
      required: [orderId, amount, method]
      properties:
        orderId:
          type: string
          maxLength: 40
        amount:
          $ref: '#/components/schemas/Money'
        method:
          $ref: '#/components/schemas/PaymentMethod'
        idempotencyKey:
          type: string
          maxLength: 120
    RefundRequest:
      type: object
      required: [amount, reason]
      properties:
        amount:
          $ref: '#/components/schemas/Money'
        reason:
          type: string
          maxLength: 200
    PaymentPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Payment'
...
    Review:
      type: object
      required: [id, productId, authorEmail, rating, text, published]
      properties:
        id:
          type: string
        productId:
          type: string
        authorEmail:
          type: string
          format: email
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        text:
          type: string
          maxLength: 4000
        published:
          type: boolean
        moderationNote:
          type: string
          maxLength: 500
        createdAt:
          type: string
          format: date-time
    ReviewCreateRequest:
      type: object
      required: [productId, authorEmail, rating, text]
      properties:
        productId:
          type: string
          maxLength: 40
        authorEmail:
          type: string
          format: email
          maxLength: 160
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        text:
          type: string
          minLength: 10
          maxLength: 4000
    ReviewPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Review'
    PromoCode:
      type: object
      required: [id, code, discountPercent, validFrom, validTo, usageLimit]
      properties:
        id:
          type: string
        code:
          type: string
          pattern: '^[A-Z0-9]{4,16}$'
        discountPercent:
          type: integer
          format: int32
          minimum: 1
          maximum: 90
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
        usageLimit:
          type: integer
          format: int32
        usedCount:
          type: integer
          format: int32
        active:
          type: boolean
    PromoCodeCreateRequest:
      type: object
      required: [code, discountPercent, validFrom, validTo, usageLimit]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9]{4,16}$'
        discountPercent:
          type: integer
          format: int32
          minimum: 1
          maximum: 90
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
        usageLimit:
          type: integer
          format: int32
          minimum: 1
          maximum: 100000
    RedemptionResult:
      type: object
      required: [code, accepted]
      properties:
        code:
          type: string
        accepted:
          type: boolean
        discountPercent:
          type: integer
          format: int32
        remainingUses:
          type: integer
          format: int32
        message:
          type: string
    PromoCodePage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/PromoCode'
...